+++ core/agent_harness/session/integration/__init__.py
@@ -1,0 +1,1 @@
+"""Per-session integration resolution, its cache, and the warmed IntegrationState."""
