+++ surfaces/interactive_shell/utils/telemetry/integration_snapshot.py
@@ -43,7 +43,7 @@ def _resolved_integrations(session: _IntegrationSession | None) -> dict[str, Any
     if session is not None and session.resolved_integrations_cache is not None:
         return session.resolved_integrations_cache
     try:
-        from core.agent_harness.session.integration_resolution import resolve_integrations
+        from core.agent_harness.session.integration.state import resolve_integrations
 
         return resolve_integrations()
     except Exception:
